% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{morrowplots}
\alias{morrowplots}
\title{Title}
\format{
\subsection{'morrowplots'}{

a description
\describe{
\item{country}{Country name}
\item{iso2,iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://doi.org/10.13012/B2IDB-7865141_V1}
}
\usage{
morrowplots
}
\description{
##add description text here
}
\keyword{datasets}
